+++ packages/noodl-editor/src/editor/src/views/NodePicker/NodePicker.hooks.ts
@@ -190,6 +190,7 @@ export function useSearchBar(
         closeAllCategories();
       }
     });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [searchTerm]);
 
   return setSearchTerm;