+++ packages/noodl-editor/src/editor/src/views/NodePicker/components/NodePickerCategory/NodePickerCategory.tsx
@@ -4,6 +4,7 @@ import React, { ReactNode, useEffect, useState } from 'react';
 
 import { NodeType } from '@noodl-constants/NodeType';
 
+import { Icon, IconName, IconSize } from '@noodl-core-ui/components/common/Icon';
 import { Collapsible } from '@noodl-core-ui/components/layout/Collapsible';
 import { Text, TextSize, TextType } from '@noodl-core-ui/components/typography/Text';
 import { Title, TitleSize, TitleVariant } from '@noodl-core-ui/components/typography/Title';
@@ -83,12 +84,13 @@ export default function NodePickerCategory({
           </Text>
         </Collapsible>
 
-        <img
-          className={classNames([
+        <Icon
+          icon={IconName.CaretRight}
+          size={IconSize.Small}
+          UNSAFE_className={classNames([
             css['Arrow'],
             isCollapsedState ? css['Arrow--is-collapsed'] : css['Arrow--is-not-collapsed']
           ])}
-          src="/assets/icons/editor/right_arrow_22.svg"
         />
       </header>